requirements.txt
asgiref==3.7.2
boto3==1.34.79
botocore==1.34.79
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
mysqlclient==2.2.1
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
